refactor: update directory structure
