+++ tomcat/src/main/java/org/apache/catalina/servlet/DefaultServlet.java
@@ -1,10 +1,10 @@
 package org.apache.catalina.servlet;
 
 import org.apache.coyote.request.HttpRequest;
-import org.apache.coyote.response.ContentType;
+import org.apache.coyote.response.responseHeader.ContentType;
 import org.apache.coyote.response.HttpResponse;
 import org.apache.coyote.response.HttpResponseGenerator;
-import org.apache.coyote.response.HttpStatus;
+import org.apache.coyote.response.responseLine.HttpStatus;
 
 public class DefaultServlet extends HttpServlet{
