+++ tomcat/src/main/java/org/apache/catalina/servlet/HttpServlet.java
@@ -3,10 +3,10 @@ package org.apache.catalina.servlet;
 import org.apache.coyote.request.HttpRequest;
 import org.apache.coyote.request.requestLine.RequestLine;
 import org.apache.coyote.request.requestLine.RequestMethod;
-import org.apache.coyote.response.ContentType;
+import org.apache.coyote.response.responseHeader.ContentType;
 import org.apache.coyote.response.HttpResponse;
 import org.apache.coyote.response.HttpResponseGenerator;
-import org.apache.coyote.response.HttpStatus;
+import org.apache.coyote.response.responseLine.HttpStatus;
 
 public abstract class HttpServlet implements Servlet {
