+++ tomcat/src/main/java/org/apache/coyote/response/responseHeader/ContentType.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.response;
+package org.apache.coyote.response.responseHeader;
 
 import java.util.Arrays;
