+++ tomcat/src/main/java/org/apache/coyote/response/responseHeader/ResponseHeader.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.response;
+package org.apache.coyote.response.responseHeader;
 
 import java.util.LinkedHashMap;
 import java.util.Map;
