+++ tomcat/src/main/java/org/apache/coyote/response/responseLine/HttpStatus.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.response;
+package org.apache.coyote.response.responseLine;
 
 public enum HttpStatus {
